Ensure press/release callbacks happen in matched pairs
A Button really ought to make the guarantee that onPress and onRelease are only ever called in matched pairs, but that wasn't previously the case. If a pulse was shorter than 10ms (or was interrupted by a long-running Squirrel callback) then the polarity could be inactive again by the time _getState was called, and you'd get a lone onRelease with no corresponding onPress. Similarly, if debouncing took longer than 10ms (or was interrupted), it could break the matched-pairs guarantee.